@@ -29,24 +29,25 @@
     "release:changelog": "npm run conventional-changelog -- -p angular -i CHANGELOG.md -s"
   },
   "dependencies": {
-    "@agm/core": "^1.0.0-beta.5",
-    "@angular/animations": "^8.0.0",
-    "@angular/cdk": "^8.0.0",
-    "@angular/common": "^8.0.0",
-    "@angular/compiler": "^8.0.0",
-    "@angular/core": "^8.0.0",
-    "@angular/forms": "^8.0.0",
-    "@angular/platform-browser": "^8.0.0",
-    "@angular/platform-browser-dynamic": "^8.0.0",
-    "@angular/router": "^8.0.0",
+    "@agm/core": "^1.1.0",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@asymmetrik/ngx-leaflet": "3.0.1",
     "@nebular/auth": "4.4.0",
     "@nebular/eva-icons": "4.4.0",
     "@nebular/security": "4.4.0",
     "@nebular/theme": "4.4.0",
     "@okta/okta-angular": "^2.0.0",
-    "@swimlane/ngx-charts": "^10.0.0",
-    "angular-oauth2-oidc": "^9.1.0",
+    "@swimlane/ngx-charts": "^10.1.0",
+    "angular-oauth2-oidc": "^9.2.0",
+    "angular-oauth2-oidc-jwks": "^9.0.0",
     "angular-tree-component": "7.2.0",
     "angular2-chartjs": "0.4.1",
     "angular2-toaster": "^7.0.0",
@@ -55,17 +56,17 @@
     "ckeditor": "4.7.3",
     "classlist.js": "1.1.20150312",
     "core-js": "2.5.1",
-    "echarts": "^4.0.2",
-    "eva-icons": "^1.1.0",
+    "echarts": "^4.7.0",
+    "eva-icons": "^1.1.3",
     "intl": "1.2.5",
     "ionicons": "2.0.1",
     "leaflet": "1.2.0",
     "nebular-icons": "1.1.0",
-    "ng2-ckeditor": "^1.2.2",
+    "ng2-ckeditor": "^1.2.7",
     "ng2-completer": "2.0.8",
     "ng2-smart-table": "1.3.5",
-    "ngx-echarts": "^4.0.1",
-    "node-sass": "^4.12.0",
+    "ngx-echarts": "^4.2.2",
+    "node-sass": "^4.13.1",
     "normalize.css": "6.0.0",
     "pace-js": "1.0.2",
     "roboto-fontface": "0.8.0",
@@ -73,25 +74,25 @@
     "rxjs-compat": "6.3.0",
     "socicon": "3.0.5",
     "tinymce": "4.5.7",
-    "tslib": "^1.9.0",
+    "tslib": "^1.11.1",
     "typeface-exo": "0.0.22",
     "web-animations-js": "github:angular/web-animations-js#release_pr208",
     "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.800.2",
-    "@angular/cli": "^8.0.2",
-    "@angular/compiler-cli": "^8.0.0",
+    "@angular-devkit/build-angular": "^0.800.6",
+    "@angular/cli": "^8.3.25",
+    "@angular/compiler-cli": "^8.2.14",
     "@angular/language-service": "8.0.0",
     "@compodoc/compodoc": "1.0.1",
-    "@fortawesome/fontawesome-free": "^5.2.0",
+    "@fortawesome/fontawesome-free": "^5.13.0",
     "@types/d3-color": "1.0.5",
-    "@types/googlemaps": "^3.30.4",
+    "@types/googlemaps": "^3.39.3",
     "@types/jasmine": "2.5.54",
     "@types/jasminewd2": "2.0.3",
     "@types/leaflet": "1.2.3",
     "@types/node": "6.0.90",
-    "codelyzer": "^5.0.1",
+    "codelyzer": "^5.2.2",
     "conventional-changelog-cli": "1.3.4",
     "husky": "0.13.3",
     "jasmine-core": "2.6.4",
@@ -107,7 +108,7 @@
     "rimraf": "2.6.1",
     "stylelint": "7.13.0",
     "ts-node": "3.2.2",
-    "tslint": "^5.7.0",
+    "tslint": "^5.20.1",
     "tslint-language-service": "^0.9.9",
     "typescript": "3.4.5"
   }
