@@ -3,8 +3,10 @@ import { NbMediaBreakpointsService, NbMenuService, NbSidebarService, NbThemeServ
 
 import { UserData } from '../../../@core/data/users';
 import { LayoutService } from '../../../@core/utils';
-import { map, takeUntil } from 'rxjs/operators';
+import { map, takeUntil, filter } from 'rxjs/operators';
 import { Subject } from 'rxjs';
+import { AuthService } from '../../../@auth/auth.service';
+import { environment } from '../../../../environments/environment';
 
 @Component({
   selector: 'ngx-header',
@@ -40,7 +42,8 @@ export class HeaderComponent implements OnInit, OnDestroy {
 
   userMenu = [ { title: 'Profile' }, { title: 'Log out' } ];
 
-  constructor(private sidebarService: NbSidebarService,
+  constructor(private authService: AuthService,
+              private sidebarService: NbSidebarService,
               private menuService: NbMenuService,
               private themeService: NbThemeService,
               private userService: UserData,
@@ -54,14 +57,35 @@ export class HeaderComponent implements OnInit, OnDestroy {
     this.userService.getUsers()
       .pipe(takeUntil(this.destroy$))
       .subscribe((users: any) => this.user = users.nick);
+    // console.debug(this.user);
 
-    const { xl } = this.breakpointService.getBreakpointsMap();
+    this.user.name = this.authService.givenName;
+
+    this.menuService.onItemClick()
+    .pipe(
+        filter(({ tag }) => tag === 'user-context-menu'),
+        map(({ item: { title } }) => title),
+    )
+    .subscribe(title => {
+        switch ( title ) {
+        case 'Profile': {
+            this.goToLink(`${environment.auth.url}`);
+            break;
+        } default: {
+            this.authService.logOut();
+            break;
+        }
+      }
+    });
+
+    // console.debug(this.breakpointService.getBreakpointsMap());
+    const { sm } = this.breakpointService.getBreakpointsMap();
     this.themeService.onMediaQueryChange()
       .pipe(
-        map(([, currentBreakpoint]) => currentBreakpoint.width < xl),
+        map(([, currentBreakpoint]) => currentBreakpoint.width < sm),
         takeUntil(this.destroy$),
       )
-      .subscribe((isLessThanXl: boolean) => this.userPictureOnly = isLessThanXl);
+      .subscribe((isLessThanSm: boolean) => this.userPictureOnly = isLessThanSm);
 
     this.themeService.onThemeChange()
       .pipe(
@@ -69,6 +93,10 @@ export class HeaderComponent implements OnInit, OnDestroy {
         takeUntil(this.destroy$),
       )
       .subscribe(themeName => this.currentTheme = themeName);
+  }
+
+  goToLink(url: string) {
+    window.open(url, '_blank');
   }
 
   ngOnDestroy() {
