@@ -66,14 +66,14 @@ export class LoginComponent extends NbLoginComponent {
 
         // Set our navigation extras object
         // that passes on our global query params and fragment
-        let navigationExtras: NavigationExtras = {
+        /* let navigationExtras: NavigationExtras = {
           queryParamsHandling: 'preserve',
           preserveFragment: true
-        };
+        }; */
 
         // Redirect the user
-        console.debug(redirect);
-        this.router.navigate([redirect], navigationExtras);
+        // console.debug(redirect);
+        this.router.navigate([redirect] /*, navigationExtras */);
       }).catch(
         err => {
           console.log('error logging in', err);
