@@ -11,6 +11,6 @@
 export const environment = {
   production: false,
 
-  auth: { url: 'https://dev-965067.oktapreview.com', issuer: 'default' },
-  clientId: '0oad8oso7zYH0gCpQ0h7',
+  auth: { url: 'https://${yourOktaDomain}', issuer: 'default' },
+  clientId: '${clientId}',
 };
